// src/net/bindings/types.ts
export const CombatEvent = __t.object("CombatEvent", {
  cell: __t.u32(),
  kind: __t.u8(),
  sourceKind: __t.u8(),
  sourceId: __t.u32(),
  targetKind: __t.u8(),
  targetId: __t.u32(),
  value: __t.u16(),
  aux: __t.u16(),
});
export type CombatEvent = __Infer<typeof CombatEvent>;